requirements.txt
# Web UI
streamlit==1.35.0
langdetect==1.0.9
requests==2.31.0

# Flask API
Flask==2.2.5
gunicorn==20.1.0

# AI関連
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
joblib==1.3.2
scikit-learn==1.4.2  # SVM系.pklモデルがある場合に必須